submission Bite 210 @ codechalleng.es

+++ 210/annotations.py
@@ -1,9 +1,9 @@
 from typing import List
-def sum_numbers(numbers: List[float]) -> int:
+def sum_numbers(numbers: List[int]) -> int:
     """Sums numbers
 
     Args:
-        numbers (List[float]): a list of numbers
+        numbers (List[int]): a list of numbers
     
     Raises:
         TypeError: if not all numeric values passed in
